Worker/pyproject.toml
[tool.poetry]
name = "notification_service"
version = "0.1.0"
description = ""
authors = ["Nozir Tukhtamurodov <nozir@9407gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
beanie = "^1.25.0"
pydantic-settings = "^2.2.1"
redis = "^5.0.1"
python-jose = "^3.3.0"
python-dotenv = "^1.0.0"
aio-pika = "^9.4.1"
Authlib = "^1.3.0"
faststream = "^0.4.7"
async-fastapi-jwt-auth = "^0.6.1"
orjson = "^3.9.10"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-bugbear = "^24.1.17"
flake8-isort = "^6.1.1"
flake8-bandit = "^4.1.1"
flake8-pytest-style = "^1.7.2"
flake8-pytest = "^1.4"
flake8-print = "^5.0.0"
flake8-commas = "^2.1.0"
flake8-broken-line = "^1.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
